reenable clockwork engine right click and move TileEngine to energy module
